# packages/api-types/openapi/v1.yaml
openapi: 3.0.3
info:
  title: EmpOps API
  version: 1.0.0
  description: Step 2 — Teams, hierarchy, and org structure (Laravel and Go parity).
servers:
  - url: http://localhost:8000/api/v1
    description: Laravel
  - url: http://localhost:8080/api/v1
    description: Go
paths:
...
  /companies/{companyId}/employees/{employeeId}/managers:
    get:
      summary: List managers for employee
      operationId: listManagers
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: Managers
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/EmployeeSummary"
    post:
      summary: Assign manager
      operationId: assignManager
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/EmployeeId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssignManagerRequest"
      responses:
        "201":
          description: Manager assigned
  /companies/{companyId}/employees/{employeeId}/managers/{managerId}:
    delete:
      summary: Unassign manager
      operationId: unassignManager
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/EmployeeId"
        - $ref: "#/components/parameters/ManagerId"
      responses:
        "200":
          description: Manager unassigned
  /companies/{companyId}/employees/{employeeId}/direct-reports:
    get:
      summary: List direct reports
      operationId: listDirectReports
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - $ref: "#/components/parameters/EmployeeId"
      responses:
        "200":
          description: Direct reports
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        type: array
                        items:
                          $ref: "#/components/schemas/EmployeeSummary"
  /companies/{companyId}/dashboard/me:
    get:
      summary: Me dashboard shell
      operationId: dashboardMe
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Dashboard shell
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/DashboardShell"
  /companies/{companyId}/dashboard/team:
    get:
      summary: Team dashboard shell
      operationId: dashboardTeam
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Dashboard shell
  /companies/{companyId}/dashboard/manager:
    get:
      summary: Manager dashboard shell
      operationId: dashboardManager
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Dashboard shell
        "403":
          description: Not a manager
  /companies/{companyId}/dashboard/hr:
    get:
      summary: HR dashboard shell
      operationId: dashboardHr
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
      responses:
        "200":
          description: Dashboard shell
        "403":
          description: Not HR/admin
  /companies/{companyId}/audit-logs:
    get:
      summary: List company audit logs
      operationId: listAuditLogs
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/CompanyId"
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Audit logs
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/ApiEnvelope"
                  - type: object
                    properties:
                      data:
                        $ref: "#/components/schemas/AuditLogPage"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    CompanyId:
      name: companyId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EmployeeId:
      name: employeeId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ManagerId:
      name: managerId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ApiEnvelope:
      type: object
      required: [success, message, data, error, timestamp]
      properties:
        success:
          type: boolean
        message:
          type: string
        data:
          nullable: true
        error:
          nullable: true
        timestamp:
          type: string
          format: date-time
...
    EmployeeSummary:
      type: object
      required: [id, first_name, last_name, email]
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        email:
          type: string
          format: email
...
    AssignManagerRequest:
      type: object
      required: [manager_id]
      properties:
        manager_id:
          type: string
          format: uuid
    DashboardShell:
      type: object
      required: [view, widgets, flags]
      properties:
        view:
          type: string
          enum: [me, team, manager, hr]
        widgets:
          type: array
          items:
            type: object
        flags:
          type: object
          properties:
            is_manager:
              type: boolean
            can_manage_hr:
              type: boolean
            is_admin:
              type: boolean
    AuditLog:
      type: object
      properties:
        id:
          type: string
        company_id:
          type: string
          format: uuid
          nullable: true
        event:
          type: string
        description:
          type: string
        subject_type:
          type: string
          nullable: true
        subject_id:
          type: string
          nullable: true
        causer_type:
          type: string
          nullable: true
        causer_id:
          type: string
          nullable: true
        properties:
          type: object
          additionalProperties: true
        created_at:
          type: string
          format: date-time
    AuditLogPage:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AuditLog"
        page:
          type: integer
        per_page:
          type: integer
        total:
          type: integer